# networking gadgets: instantiate ebpf prog once per netns