--- a/pkg/gadget-collection/gadgets/trace/network/gadget.go
+++ b/pkg/gadget-collection/gadgets/trace/network/gadget.go
@@ -20,7 +20,6 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	gadgetv1alpha1 "github.com/inspektor-gadget/inspektor-gadget/pkg/apis/gadget/v1alpha1"
-	containercollection "github.com/inspektor-gadget/inspektor-gadget/pkg/container-collection"
 	"github.com/inspektor-gadget/inspektor-gadget/pkg/container-collection/networktracer"
 	"github.com/inspektor-gadget/inspektor-gadget/pkg/gadget-collection/gadgets"
 	netTracer "github.com/inspektor-gadget/inspektor-gadget/pkg/gadgets/trace/network/tracer"
@@ -125,12 +124,11 @@ func (t *Trace) Start(trace *gadgetv1alpha1.Trace) {
 		inst.PreGadgetRun()
 	}
 
-	eventCallback := func(container *containercollection.Container, event *netTypes.Event) {
-		// Enrich event with data from container
+	eventCallback := func(event *netTypes.Event) {
+		// Enrich event but only with the fields required for the advise network-policy gadget.
 		event.Node = trace.Spec.Node
-		if !container.HostNetwork {
-			event.Namespace = container.Namespace
-			event.Pod = container.Podname
+		if t.helpers != nil {
+			t.helpers.EnrichByNetNs(&event.CommonData, event.WithNetNsID.NetNsID)
 		}
 
 		// Use KubeIPResolver to enrich event based on Namespace/Pod and IP.
@@ -140,13 +138,13 @@ func (t *Trace) Start(trace *gadgetv1alpha1.Trace) {
 
 		t.publishEvent(trace, event)
 	}
+	t.tracer.SetEventHandler(eventCallback)
 
 	config := &networktracer.ConnectToContainerCollectionConfig[netTypes.Event]{
-		Tracer:        t.tracer,
-		Resolver:      t.helpers,
-		Selector:      *gadgets.ContainerSelectorFromContainerFilter(trace.Spec.Filter),
-		EventCallback: eventCallback,
-		Base:          netTypes.Base,
+		Tracer:   t.tracer,
+		Resolver: t.helpers,
+		Selector: *gadgets.ContainerSelectorFromContainerFilter(trace.Spec.Filter),
+		Base:     netTypes.Base,
 	}
 	t.conn, err = networktracer.ConnectToContainerCollection(config)
 	if err != nil {