--- a/pkg/gadget-collection/gadgets/trace/sni/gadget.go
+++ b/pkg/gadget-collection/gadgets/trace/sni/gadget.go
@@ -20,7 +20,6 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	gadgetv1alpha1 "github.com/inspektor-gadget/inspektor-gadget/pkg/apis/gadget/v1alpha1"
-	containercollection "github.com/inspektor-gadget/inspektor-gadget/pkg/container-collection"
 	"github.com/inspektor-gadget/inspektor-gadget/pkg/container-collection/networktracer"
 	"github.com/inspektor-gadget/inspektor-gadget/pkg/gadget-collection/gadgets"
 	sniTracer "github.com/inspektor-gadget/inspektor-gadget/pkg/gadgets/trace/sni/tracer"
@@ -107,30 +106,23 @@ func (t *Trace) Start(trace *gadgetv1alpha1.Trace) {
 		return
 	}
 
+	eventCallback := func(event *sniTypes.Event) {
+		t.publishEvent(trace, event)
+	}
+
 	var err error
 	t.tracer, err = sniTracer.NewTracer()
 	if err != nil {
 		trace.Status.OperationError = fmt.Sprintf("Failed to start sni tracer: %s", err)
 		return
 	}
-
-	eventCallback := func(container *containercollection.Container, event *sniTypes.Event) {
-		// Enrich event with data from container
-		event.Node = trace.Spec.Node
-		if !container.HostNetwork {
-			event.Namespace = container.Namespace
-			event.Pod = container.Podname
-		}
-
-		t.publishEvent(trace, event)
-	}
+	t.tracer.SetEventHandler(eventCallback)
 
 	config := &networktracer.ConnectToContainerCollectionConfig[sniTypes.Event]{
-		Tracer:        t.tracer,
-		Resolver:      t.helpers,
-		Selector:      *gadgets.ContainerSelectorFromContainerFilter(trace.Spec.Filter),
-		EventCallback: eventCallback,
-		Base:          sniTypes.Base,
+		Tracer:   t.tracer,
+		Resolver: t.helpers,
+		Selector: *gadgets.ContainerSelectorFromContainerFilter(trace.Spec.Filter),
+		Base:     sniTypes.Base,
 	}
 	t.conn, err = networktracer.ConnectToContainerCollection(config)
 	if err != nil {