--- a/pkg/gadgets/internal/networktracer/tracer.go
+++ b/pkg/gadgets/internal/networktracer/tracer.go
@@ -19,6 +19,7 @@ import (
 	"fmt"
 	"os"
 	"syscall"
+	"unsafe"
 
 	"github.com/cilium/ebpf"
 	"github.com/cilium/ebpf/perf"
@@ -33,13 +34,14 @@ import (
 	"github.com/inspektor-gadget/inspektor-gadget/pkg/types"
 )
 
+//go:generate bash -c "source ./clangosflags.sh; go run github.com/cilium/ebpf/cmd/bpf2go -target bpfel -cc clang dispatcher ./bpf/dispatcher.bpf.c -- $CLANG_OS_FLAGS -I./bpf/ -I../socketenricher/bpf"
+
 const (
 	SocketsMapName = "sockets"
 )
 
 type attachment struct {
-	collection *ebpf.Collection
-	perfRd     *perf.Reader
+	dispatcherObjs dispatcherObjects
 
 	sockFd int
 
@@ -51,14 +53,24 @@ type attachment struct {
 	users map[uint32]struct{}
 }
 
-func newAttachment(
+type Tracer[Event any] struct {
+	socketEnricher *socketenricher.SocketEnricher
+	collection     *ebpf.Collection
+	prog           *ebpf.Program
+	perfRd         *perf.Reader
+
+	// key: network namespace inode number
+	// value: Tracelet
+	attachments map[uint64]*attachment
+
+	baseEvent    func(ev types.Event) *Event
+	processEvent func(rawSample []byte, netns uint64) (*Event, error)
+	eventHandler func(ev *Event)
+}
+
+func (t *Tracer[Event]) newAttachment(
 	pid uint32,
 	netns uint64,
-	socketEnricher *socketenricher.SocketEnricher,
-	spec *ebpf.CollectionSpec,
-	bpfProgName string,
-	bpfPerfMapName string,
-	bpfSocketAttach int,
 ) (_ *attachment, err error) {
 	a := &attachment{
 		sockFd: -1,
@@ -66,50 +78,34 @@ func newAttachment(
 	}
 	defer func() {
 		if err != nil {
-			if a.perfRd != nil {
-				a.perfRd.Close()
-			}
 			if a.sockFd != -1 {
 				unix.Close(a.sockFd)
 			}
-			if a.collection != nil {
-				a.collection.Close()
-			}
+			a.dispatcherObjs.Close()
 		}
 	}()
 
-	spec = spec.Copy()
-
-	var opts ebpf.CollectionOptions
-
-	if socketEnricher != nil {
-		u32netns := uint32(netns)
-		consts := map[string]interface{}{
-			"current_netns": u32netns,
-		}
-
-		if err := spec.RewriteConstants(consts); err != nil {
-			return nil, fmt.Errorf("rewriting constants while attaching to pid %d: %w", pid, err)
-		}
-
-		mapReplacements := map[string]*ebpf.Map{}
-		mapReplacements[SocketsMapName] = socketEnricher.SocketsMap()
-		opts.MapReplacements = mapReplacements
-	}
-
-	a.collection, err = ebpf.NewCollectionWithOptions(spec, opts)
+	dispatcherSpec, err := loadDispatcher()
 	if err != nil {
-		return nil, fmt.Errorf("creating BPF collection: %w", err)
+		return nil, err
 	}
 
-	a.perfRd, err = perf.NewReader(a.collection.Maps[bpfPerfMapName], gadgets.PerfBufferPages*os.Getpagesize())
-	if err != nil {
-		return nil, fmt.Errorf("getting a perf reader: %w", err)
+	u32netns := uint32(netns)
+	consts := map[string]interface{}{
+		"current_netns": u32netns,
 	}
-
-	prog, ok := a.collection.Programs[bpfProgName]
-	if !ok {
-		return nil, fmt.Errorf("BPF program %q not found", bpfProgName)
+	if err := dispatcherSpec.RewriteConstants(consts); err != nil {
+		return nil, fmt.Errorf("RewriteConstants while attaching to pid %d: %w", pid, err)
+	}
+	dispatcherSpec.Maps["tail_call"].Contents = []ebpf.MapKV{
+		{
+			Key:   uint32(0),
+			Value: uint32(t.prog.FD()),
+		},
+	}
+	opts := ebpf.CollectionOptions{}
+	if err = dispatcherSpec.LoadAndAssign(&a.dispatcherObjs, &opts); err != nil {
+		return nil, fmt.Errorf("loading ebpf program: %w", err)
 	}
 
 	a.sockFd, err = rawsock.OpenRawSock(pid)
@@ -117,48 +113,47 @@ func newAttachment(
 		return nil, fmt.Errorf("opening raw socket: %w", err)
 	}
 
-	if err := syscall.SetsockoptInt(a.sockFd, syscall.SOL_SOCKET, bpfSocketAttach, prog.FD()); err != nil {
+	if err := syscall.SetsockoptInt(a.sockFd, syscall.SOL_SOCKET, unix.SO_ATTACH_BPF, a.dispatcherObjs.IgNetDisp.FD()); err != nil {
 		return nil, fmt.Errorf("attaching BPF program: %w", err)
 	}
-
 	return a, nil
-}
-
-type Tracer[Event any] struct {
-	socketEnricher *socketenricher.SocketEnricher
-	spec           *ebpf.CollectionSpec
-
-	// key: network namespace inode number
-	// value: Tracelet
-	attachments map[uint64]*attachment
-
-	bpfProgName     string
-	bpfPerfMapName  string
-	bpfSocketAttach int
-
-	baseEvent    func(ev types.Event) *Event
-	processEvent func(rawSample []byte, netns uint64) (*Event, error)
-
-	eventHandler func(ev *Event)
 }
 
 func NewTracer[Event any](
 	spec *ebpf.CollectionSpec,
 	bpfProgName string,
 	bpfPerfMapName string,
-	bpfSocketAttach int,
 	baseEvent func(ev types.Event) *Event,
 	processEvent func(rawSample []byte, netns uint64) (*Event, error),
-) (*Tracer[Event], error) {
+) (_ *Tracer[Event], err error) {
 	gadgets.FixBpfKtimeGetBootNs(spec.Programs)
 
-	var socketEnricher *socketenricher.SocketEnricher
-	var err error
+	t := &Tracer[Event]{
+		attachments:  make(map[uint64]*attachment),
+		baseEvent:    baseEvent,
+		processEvent: processEvent,
+	}
+
+	defer func() {
+		if err != nil {
+			if t.perfRd != nil {
+				t.perfRd.Close()
+			}
+			if t.collection != nil {
+				t.collection.Close()
+			}
+			if t.socketEnricher != nil {
+				t.socketEnricher.Close()
+			}
+		}
+	}()
+
+	var opts ebpf.CollectionOptions
 
 	// Only create socket enricher if this is used by the tracer
 	for _, m := range spec.Maps {
 		if m.Name == SocketsMapName {
-			socketEnricher, err = socketenricher.NewSocketEnricher()
+			t.socketEnricher, err = socketenricher.NewSocketEnricher()
 			if err != nil {
 				// Non fatal: support kernels without BTF
 				log.Errorf("creating socket enricher: %s", err)
@@ -167,19 +162,32 @@ func NewTracer[Event any](
 		}
 	}
 
-	return &Tracer[Event]{
-		socketEnricher:  socketEnricher,
-		spec:            spec,
-		attachments:     make(map[uint64]*attachment),
-		bpfProgName:     bpfProgName,
-		bpfPerfMapName:  bpfPerfMapName,
-		bpfSocketAttach: bpfSocketAttach,
-		baseEvent:       baseEvent,
-		processEvent:    processEvent,
-	}, nil
+	if t.socketEnricher != nil {
+		mapReplacements := map[string]*ebpf.Map{}
+		mapReplacements[SocketsMapName] = t.socketEnricher.SocketsMap()
+		opts.MapReplacements = mapReplacements
+	}
+
+	t.collection, err = ebpf.NewCollectionWithOptions(spec, opts)
+	if err != nil {
+		return nil, fmt.Errorf("creating BPF collection: %w", err)
+	}
+
+	t.perfRd, err = perf.NewReader(t.collection.Maps[bpfPerfMapName], gadgets.PerfBufferPages*os.Getpagesize())
+	if err != nil {
+		return nil, fmt.Errorf("getting a perf reader: %w", err)
+	}
+
+	var ok bool
+	t.prog, ok = t.collection.Programs[bpfProgName]
+	if !ok {
+		return nil, fmt.Errorf("BPF program %q not found", bpfProgName)
+	}
+
+	return t, nil
 }
 
-func (t *Tracer[Event]) Attach(pid uint32, eventCallback func(*Event)) error {
+func (t *Tracer[Event]) Attach(pid uint32) error {
 	netns, err := containerutils.GetNetNs(int(pid))
 	if err != nil {
 		return fmt.Errorf("getting network namespace of pid %d: %w", pid, err)
@@ -189,28 +197,42 @@ func (t *Tracer[Event]) Attach(pid uint32, eventCallback func(*Event)) error {
 		return nil
 	}
 
-	a, err := newAttachment(pid, netns, t.socketEnricher, t.spec,
-		t.bpfProgName, t.bpfPerfMapName, t.bpfSocketAttach)
+	a, err := t.newAttachment(pid, netns)
 	if err != nil {
 		return fmt.Errorf("creating network tracer attachment for pid %d: %w", pid, err)
 	}
 	t.attachments[netns] = a
 
-	go t.listen(netns, a.perfRd, t.baseEvent, t.processEvent, eventCallback)
-
 	return nil
 }
 
 func (t *Tracer[Event]) SetEventHandler(handler any) {
+	if t.eventHandler != nil {
+		panic("handler already set")
+	}
+
 	nh, ok := handler.(func(ev *Event))
 	if !ok {
 		panic("event handler invalid")
 	}
 	t.eventHandler = nh
+	go t.listen(t.perfRd, t.baseEvent, t.processEvent, t.eventHandler)
+}
+
+// EventCallback provides support for legacy pkg/gadget-collection
+func (t *Tracer[Event]) EventCallback(event any) {
+	e, ok := event.(*Event)
+	if !ok {
+		panic("event handler argument invalid")
+	}
+	if t.eventHandler == nil {
+		return
+	}
+	t.eventHandler(e)
 }
 
 func (t *Tracer[Event]) AttachContainer(container *containercollection.Container) error {
-	return t.Attach(container.Pid, t.eventHandler)
+	return t.Attach(container.Pid)
 }
 
 func (t *Tracer[Event]) DetachContainer(container *containercollection.Container) error {
@@ -218,7 +240,6 @@ func (t *Tracer[Event]) DetachContainer(container *containercollection.Container
 }
 
 func (t *Tracer[Event]) listen(
-	netns uint64,
 	rd *perf.Reader,
 	baseEvent func(ev types.Event) *Event,
 	processEvent func(rawSample []byte, netns uint64) (*Event, error),
@@ -231,18 +252,25 @@ func (t *Tracer[Event]) listen(
 				return
 			}
 
-			msg := fmt.Sprintf("Error reading perf ring buffer (%d): %s", netns, err)
+			msg := fmt.Sprintf("Error reading perf ring buffer: %s", err)
 			eventCallback(baseEvent(types.Err(msg)))
 			return
 		}
 
 		if record.LostSamples != 0 {
-			msg := fmt.Sprintf("lost %d samples (%d)", record.LostSamples, netns)
+			msg := fmt.Sprintf("lost %d samples", record.LostSamples)
 			eventCallback(baseEvent(types.Warn(msg)))
 			continue
 		}
 
-		event, err := processEvent(record.RawSample, netns)
+		if len(record.RawSample) < 4 {
+			eventCallback(baseEvent(types.Err("record too small")))
+			continue
+		}
+
+		// all networking gadgets have netns as first field
+		netns := *(*uint32)(unsafe.Pointer(&record.RawSample[0]))
+		event, err := processEvent(record.RawSample, uint64(netns))
 		if err != nil {
 			eventCallback(baseEvent(types.Err(err.Error())))
 			continue
@@ -255,9 +283,8 @@ func (t *Tracer[Event]) listen(
 }
 
 func (t *Tracer[Event]) releaseAttachment(netns uint64, a *attachment) {
-	a.perfRd.Close()
 	unix.Close(a.sockFd)
-	a.collection.Close()
+	a.dispatcherObjs.Close()
 	delete(t.attachments, netns)
 }
 
@@ -275,6 +302,8 @@ func (t *Tracer[Event]) Detach(pid uint32) error {
 }
 
 func (t *Tracer[Event]) Close() {
+	t.perfRd.Close()
+	t.collection.Close()
 	for key, l := range t.attachments {
 		t.releaseAttachment(key, l)
 	}