--- a/pkg/gadgets/trace/dns/tracer/bpf/dns-common.h
+++ b/pkg/gadgets/trace/dns/tracer/bpf/dns-common.h
@@ -13,6 +13,9 @@
 #define MAX_ADDR_ANSWERS 8
 
 struct event_t {
+	// Keep netns at the top: networktracer depends on it
+	__u32 netns;
+
 	__u64 timestamp;
 	__u64 mount_ns_id;
 	__u32 pid;