--- a/pkg/gadgets/trace/dns/tracer/dns_bpfel.go
+++ b/pkg/gadgets/trace/dns/tracer/dns_bpfel.go
@@ -13,6 +13,8 @@ import (
 )
 
 type dnsEventT struct {
+	Netns       uint32
+	_           [4]byte
 	Timestamp   uint64
 	MountNsId   uint64
 	Pid         uint32