--- a/pkg/gadgets/trace/sni/tracer/snisnoop_bpfel.go
+++ b/pkg/gadgets/trace/sni/tracer/snisnoop_bpfel.go
@@ -13,6 +13,8 @@ import (
 )
 
 type snisnoopEventT struct {
+	Netns     uint32
+	_         [4]byte
 	MountNsId uint64
 	Pid       uint32
 	Tid       uint32