… pkg/gadgets/internal/networktracer/clangosflags.sh
… pkg/gadgets/internal/networktracer/dispatcher_bpfel.go
… pkg/gadgets/internal/networktracer/dispatcher_bpfel.o
… pkg/gadgets/trace/dns/tracer/dns_bpfel.o
… pkg/gadgets/trace/network/tracer/clangosflags.sh
… pkg/gadgets/trace/network/tracer/network_bpfel.o
… pkg/gadgets/trace/sni/tracer/clangosflags.sh
… pkg/gadgets/trace/sni/tracer/snisnoop_bpfel.o
… pkg/gadgets/trace/tcpdrop/tracer/tcpdrop_bpfel_arm64.o
… pkg/gadgets/trace/tcpdrop/tracer/tcpdrop_bpfel_x86.o
… pkg/gadgets/trace/tcpretrans/tracer/tcpretrans_bpfel_arm64.o
… pkg/gadgets/trace/tcpretrans/tracer/tcpretrans_bpfel_x86.o