# networking gadgets: instantiate ebpf prog once per netns

In the tests with 100 containers on a node, the networking gadgets were
sometimes failing by timeout because they were too slow. Profiling
showed that most of the time is spent in the ebpf verifier. That is
because networking gadgets load the ebpf program once per network
namespace.

This patch refactors networking gadgets to use only one instance of ebpf
gadget. It is used for all network namespace. Unfortunately, we cannot
attach the ebpf program directly on the raw socket for each network
namespace because we need to know the netns in the ebpf program to get
the enrichment working properly.

So we have a new, very small ebpf program named 'dispatcher',
instantiated for each netns. Each instance is attached to a raw socket
on that netns. Each instance knows the current_netns then doing a tail
call on the singleton ebpf program. Since we cannot pass parameters in
tail calls, we use the skb->cb[] buffer to pass the context.

Since the dispatcher is very small, the ebpf verifier is still fast
enough, even with 100 instances.

Thanks @mauriciovasquezbernal for the idea to use skb->cb[] and tail calls.

The [documentation for skb->cb[]](http://vger.kernel.org/~davem/skb.html) says:
> This is the SKB control block. It is an opaque storage area usable by protocols, and even some drivers, to store private per-packet information. TCP uses this, for example, to store sequence numbers and retransmission state for the frame.

In the context of ebpf programs running on socket filters, the skb given to the ebpf program is a [skb clone](http://vger.kernel.org/~davem/skb.html):

> In order to make quick references to SKB data, Linux has the concept of SKB clones. When a clone of an SKB is made, all of the 'struct sk_buff' structure members of the clone are private to the clone. The data, however, is shared between the primary SKB and it's clone. When an SKB is cloned, the 'cloned' field will be set in both the primary and clone SKB. Otherwise is will be zero.

So reading/writing on skb->cb[] from the socket filter will not interfere with the networking stack.

## How to use

Run the benchmarks:
```
IG_BENCHMARKS_GADGET_REGEX='trace-[ds]n[si]' go test -exec 'sudo -E' -bench=. -run=Benchmark ./internal/benchmarks/... -count 8 |tee main.bench
IG_BENCHMARKS_GADGET_REGEX='trace-[ds]n[si]' go test -exec 'sudo -E' -bench=. -run=Benchmark ./internal/benchmarks/... -count 8 |tee patched.bench
```

Check the results:
```
$ benchstat main.bench patched.bench
goos: linux
goarch: amd64
pkg: github.com/inspektor-gadget/inspektor-gadget/internal/benchmarks
cpu: Intel(R) Core(TM) i7-6500U CPU @ 2.50GHz
                                     │ main.bench  │           patched.bench           │
                                     │   sec/op    │   sec/op    vs base               │
AllGadgetsManyContainers/trace-dns-4   2.945 ± 11%   1.584 ± 8%  -46.21% (p=0.000 n=8)
AllGadgetsManyContainers/trace-sni-4   4.735 ± 11%   1.523 ± 9%  -67.82% (p=0.000 n=8)
geomean                                3.734         1.553       -58.40%
```

Benchmarks showed a 58% performance improvement when starting a
networking gadget with 100 containers.

## Testing done

## TODO

- [x] Get https://github.com/inspektor-gadget/inspektor-gadget/pull/1419 merged first
- [x] Test the 3 networking gadgets
  - [x] dns
  - [x] sni
  - [x] network
- [x] Test with different frontends
  - [x] kubectl-gadget
  - [x] ig
- [x] Test the enrichment (node, pod, namespace...)
  - [x] in the main events
  - [x] in the debug events ("tracer attached" messages)

-----

This PR is part of https://github.com/inspektor-gadget/inspektor-gadget/issues/1394

Fixes https://github.com/inspektor-gadget/inspektor-gadget/issues/1612