@alban 2023-03-13
I ran more benchmarks with 0, 1, 10 and 100 containers:

```
IG_BENCHMARKS_GADGET_REGEX='[sd]n[is]' go test -exec 'sudo -E' -bench=. -run=Benchmark ./internal/benchmarks/... -count 8 | tee main.bench 
IG_BENCHMARKS_GADGET_REGEX='[sd]n[is]' go test -exec 'sudo -E' -bench=. -run=Benchmark ./internal/benchmarks/... -count 8 | tee patched.bench 

$ benchstat main.bench patched.bench
goos: linux
goarch: amd64
pkg: github.com/inspektor-gadget/inspektor-gadget/internal/benchmarks
cpu: Intel(R) Core(TM) i7-6500U CPU @ 2.50GHz
                                                  │  main.bench   │            patched.bench            │
                                                  │    sec/op     │    sec/op     vs base               │
AllGadgetsWithContainers/container0/trace-dns-4      854.2m ± 76%   797.1m ± 16%        ~ (p=0.721 n=8)
AllGadgetsWithContainers/container0/trace-sni-4      923.2m ± 22%   927.8m ± 16%        ~ (p=0.721 n=8)
AllGadgetsWithContainers/container1/trace-dns-4      878.3m ± 11%   806.2m ±  7%        ~ (p=0.279 n=8)
AllGadgetsWithContainers/container1/trace-sni-4      839.9m ± 27%   834.6m ± 21%        ~ (p=0.878 n=8)
AllGadgetsWithContainers/container10/trace-dns-4    1072.8m ± 16%   881.4m ± 22%  -17.84% (p=0.028 n=8)
AllGadgetsWithContainers/container10/trace-sni-4    1388.1m ± 52%   932.0m ± 11%  -32.86% (p=0.000 n=8)
AllGadgetsWithContainers/container100/trace-dns-4     3.345 ± 16%    1.511 ± 10%  -54.83% (p=0.000 n=8)
AllGadgetsWithContainers/container100/trace-sni-4     4.693 ± 14%    1.516 ± 12%  -67.70% (p=0.000 n=8)
geomean                                               1.386         991.8m        -28.43%
```