@alban 2023-03-15
With the current state of this PR, the container enrichment does not work everywhere:
- **ig**
  - ✔️ **dns**
  - ✔️ **sni**
  - ❌ **network**: see explanation below
- ❌ **kubectl-gadget**: it does not use the new operators. I didn't investigate more; I am wondering whether to wait for #1396 to fix it.

Currently, `ig trace network -o json` shows events with netnsid=0. This is because the trace-network gadget does not use the `internal/networktracer` package like dns and sni, so it does not use the new dispatcher, so `skb->cb[0]` is not set. I see two choices to fix it:
1. Update the trace-network gadget to use `internal/networktracer`, so it uses the dispatcher, benefit from the perf improvements.
2. Continue to use `const volatile u64 container_netns` in `pkg/gadgets/trace/network/tracer/bpf/graph.c`.

I am thinking choice 1 is better, because we will need it eventually to have good performances in the network-policy advisor. Maybe this could be done before this PR? Then this PR will be a bit smaller to review.