@mauriciovasquezbernal 2023-03-15
> I am thinking choice 1 is better

I agree. That'd be great if we're able to uniform those networking gadgets even more. It worries me a little bit that we're creating some dependencies between PRs but I think it's the only way to keep this organized and avoid a lot of conflicts.